requirements.txt
graphql-core>=3